refactor(진성진): 컴포넌트 폴더 일부 이동
