+++ src/components/index.ts
@@ -1,8 +1,4 @@
-import MainSection from "./MainSection/MainSection";
-import WideSection from "./WideSection/WideSection";
-
-import ImageField from "./ImageField/ImageField";
-import TagField from "./TagField/TagField";
+import Field from "./Field/Field";
 import InputField from "./InputField/InputField";
 import Button from "./Button/Button";
 import ConvenientSigninBox from "./ConvenientSigninBox/ConvenientSigninBox";
@@ -15,19 +11,11 @@ import Input from "./Input/Input";
 import Select from "./Select/Select";
 import Pagination from "./Pagination/Pagination";
 
-import ItemInfo from "./ItemInfo/ItemInfo";
-import ItemComment from "./ItemComment/ItemComment";
-import CommentForm from "./CommentForm/CommentForm";
-import CommentList from "./CommentList/CommentList";
-
 import Profile from "./Profile/Profile";
 import Dropdown from "./Dropdown/Dropdown";
 
 export {
-  MainSection,
-  WideSection,
-  TagField,
-  ImageField,
+  Field,
   InputField,
   Button,
   ConvenientSigninBox,
@@ -37,10 +25,6 @@ export {
   Select,
   Input,
   Pagination,
-  ItemInfo,
-  ItemComment,
-  CommentForm,
-  CommentList,
   Profile,
   Dropdown,
 };
