+++ src/pages/AddItemPage/components/index.ts
@@ -1,0 +1,4 @@
+import TagField from "./TagField/TagField";
+import ImageField from "./ImageField/ImageField";
+
+export { TagField, ImageField };
