+++ src/pages/HomePage/components/MainSection/MainSection.module.css
