+++ src/pages/HomePage/components/WideSection/WideSection.module.css
