+++ src/pages/HomePage/components/index.ts
@@ -1,0 +1,4 @@
+import MainSection from "./MainSection/MainSection";
+import WideSection from "./WideSection/WideSection";
+
+export { MainSection, WideSection };
