+++ src/pages/ItemPage/components/index.ts
@@ -1,0 +1,6 @@
+import ItemInfo from "./ItemInfo/ItemInfo";
+import ItemComment from "./ItemComment/ItemComment";
+import CommentList from "./CommentList/CommentList";
+import CommentForm from "./CommentForm/CommentForm";
+
+export { ItemInfo, ItemComment, CommentList, CommentForm };
